[FIX] website,website_livechat : restore chat request feature
since 6521ab8b223e2b688ccf809dec9d1b0cc2d4c2cd,
a chat request could not be sent to a visitor as the website was not
set anymore on the visitor and the mail_channel could then not be
assigned to a livechat_channel (using the website_id).

This commit restore the feature using the new temporary model
website.visitor.lastconnection.

Task ID: 2056080
PR #36290