[FIX] sql_db: retrieve correct environment
When creating a new database from the database manager, with no open
cursor, the wrong env was retrieved during the cr.commit() call.

2019-09-09 18:16:07,454 19886 ERROR None odoo.service.db: CREATE DATABASE failed:
Traceback (most recent call last):
  File "/home/odoo/odoo/service/db.py", line 60, in _initialize_db
    cr.commit()
  File "/home/odoo/odoo/sql_db.py", line 165, in wrapper
    return f(self, *args, **kwargs)
  File "/home/odoo/odoo/sql_db.py", line 389, in commit
    env = get_env(currentframe(), 2)
  File "/home/odoo/odoo/sql_db.py", line 72, in get_env
    env = getattr(frame.f_locals.get('self'), 'env', SENTINEL)
  File "/home/odoo/odoo/http.py", line 268, in env
    self._env = odoo.api.Environment(self.cr, self.uid, self.context)
  File "/home/odoo/odoo/http.py", line 238, in cr
    raise RuntimeError('request not bound to a database')

get_env method checks every frame to find a variable called 'env'

As none was found, it used the one of the WebRequest object.
In a WebRequest, env is a property that will try to create a new
Environment, which fails as no database is available (yet).

At 98ac5d09fb8 the iteration on Environment.envs was removed as envs
is a weakset and it introduced potential errors where the content was
garbage-collected during the iteration.

Go back to the simpler iteration but convert the set to list to avoid
garbage-collection issues.

Fixes odoo/odoo#36588

closes odoo/odoo#37005

Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>
Co-authored-by: Raphael Collet <rco@odoo.com>