[FIX] sale_crm: correctly handle customer option when converting opportunity
When converting an opportunity to a quotation, it will give 3 options where
the last option is "Do not link to a customer". When clicked it has the
same behavior than "Create a new customer".

Followup of ac1a6434ff1800b351633937cdf2730d7b2893bb that added the
selection entry but did not implement its behavior.

This commit fixes that behavior. Now it opens a quotation without linked
customer.

Task ID 2065014

closes odoo/odoo#37026

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>