[IMP] website, website_sale: Improve website tracking
The update of visit_count now uses website_track instead of
website_visitor_lastconnection for less complex query and better
precision.

The name of a visitor is now changed only if it has no partner linked
and the # is removed

The cookie was not reset when a new partner connects or if a partner
disconnects. Now a check is performed to ensure the cookie belongs to
the partner/public user.

Now website visitors that log in will be added to the previous
website visitor of the partner. This way we have a full history and only
one visitor per partner.

Now a partner is able to delete a recently viewed product from his
recently viewed product list.

task-2072877