[FIX] mass mailing: fix some mailing form view glitches
In the email/sms marketing form (edition), only on debug mode, we can get
a js traceback (bootstrap tooltip can't be show for a invisible element,
the domain field in this case) when we change the recipient model from Mailing
List -> a other model -> Mailing List. We avoid the traceback by putting
invisible a parent div of the domain fields instead of the domain field
itself.

In the email/sms marketing form (edition), we can create a email/sms marketing
without any recipient model which cause a issue on the kanban view (blank
space). Also behind the scene the mass mailing contact model is used. To
clarify and fix the issue in kanban view, the recipient model is now required
in the form view.

In the email/sms marketing form (debug mode only), we can create a email/sms
marketing with "allow A/B testing" checked without a campaign marketing
(not logical), which lead to a traceback when the email are sent. To fix
this issue, we move the "Allow A/B testing" field in the group form called
"marketing" and when it is checked the campaign marketing field is mandatory.

Task ID 2058738 (mass mailing internal testing)
PR #37185