[FIX] hr_holidays: Set leave manager if empty
Currently, when an employee's manager (parent_id) changes,
the leave manager (leave_manager_id) is set to the new manager
if it was not manually changed (leave_manager_id != parent_id.user_id).

Please, note that `parent_id` is an `hr.employee` while `leave_manager_id`
is a `res.users`. Hence if `leave_manager_id` is not set, we have
`leave_manager_id == False`.
Next time we change `parent_id` (employee with a user)
we have `False != parent_id.user_id` and the leave_manager_id
is not set.