[FIX] l10n_fr_pos_cert: avoid editing orders
Once an order is created in the french certification, we cannot modify
it wihout traces, so we've disabled the edtition of the order lines in
the frontend, disable the price changing, and the deletion of orders.

So when we want to decrease the quantity, a line with negative quantity
should be done.

TASK-ID: 47990